refactor: adjust files' path
