--- src/data/newsType.js
@@ -1,14 +1,0 @@
-export const newsType = [
-  {
-    msg: '注意',
-    color: 'warning'
-  },
-  {
-    msg: '部分封閉',
-    color: 'secondary'
-  },
-  {
-    msg: '暫停開放',
-    color: 'danger'
-  }
-]
