+++ src/utils/trailInfoUtils.js
@@ -43,3 +43,18 @@ export const extendedTrailInfo = {
     { type: 'TR_ADMIN_PHONE', name: '管轄單位聯繫' }
   ]
 }
+
+export const newsType = [
+  {
+    msg: '注意',
+    color: 'warning'
+  },
+  {
+    msg: '部分封閉',
+    color: 'secondary'
+  },
+  {
+    msg: '暫停開放',
+    color: 'danger'
+  }
+]
